[doyto-query-mongodb] fix type in QuantityView

+++ doyto-query-mongodb/src/test/java/win/doyto/query/mongodb/test/aggregate/QuantityView.java
@@ -35,6 +35,6 @@ public class QuantityView {
     private Double avgQty;
     private Integer firstQty;
     private Integer lastQty;
-    private Integer stdDevPopQty;
-    private Integer stdDevSampQty;
+    private Double stdDevPopQty;
+    private Double stdDevSampQty;
 }
